Fixing #462
When switching panels the old panel is now destroyed before the new panel is created.
In addition 'beforedestroy' handlers are modified so added map drawcontrols are removed now. Also extra 'parenthide' handlers are added to deactivate the current drawcontrol when the whole form is closed.

Solving this issue is commissioned by Gemeente Nieuwegein, The Netherlands.